Renamed vobject to vobjectx
all occurrences replaced
