+++ vobjectx/__about__.py
