+++ vobjectx/exceptions.py
