+++ vobjectx/helper/funcs.py
