+++ vobjectx/patterns.py
